+++ SignNow.Net/Interfaces/ISignInvite.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using SignNow.Net.Model;
@@ -63,5 +61,13 @@ namespace SignNow.Net.Interfaces
         /// <param name="cancellationToken">Propagates notification that operations should be canceled.</param>
         /// <returns></returns>
         Task CancelEmbeddedInviteAsync(string documentId, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Resends an invite to sign a document.
+        /// </summary>
+        /// <param name="fieldInviteId">The role identity to resend the invitation to.</param>
+        /// <param name="cancellationToken">Propagates notification that operations should be canceled.</param>
+        /// <returns></returns>
+        Task ResendEmailInviteAsync(string fieldInviteId, CancellationToken cancellationToken = default);
     }
 }